Create My Store Sales
I have a lot of data for my resale store located in Buffalo, NY. We buy from the public daily, so we have to understand our inventory on a deep analytical level. In order to inform my buyers as to what's selling well and what we should continue to purchase from our customers, I use SQL to organize and analyze our sales data monthly. I then present this data to our team, we formulate additional questions about it, and I continue to analyze the data and repeat this process.